requirements.txt
# Sugar Dependencies - Minimal and focused
aiosqlite>=0.19.0
pyyaml>=6.0
watchdog>=3.0.0
gitpython>=3.1.40
pygithub>=1.59.0
structlog>=23.0.0
click>=8.1.0
python-dotenv>=1.0.0